voice-backend/requirements.txt
flask 
flask-cors
TTS
